refactor: separate BpSideBarItem in a new file
